Check against DBA
This check works for t5, which is a breakout strategy to go in the
direction of the breakout evertime the price breaches a 100d high or
low.  This works successfully over the last 5y of SP500, and now I've
also tested this successfully over the last 5y of DBA.  The DBA test
is better since the SP500 has been a pretty solid bull market the
whole time.  The DBA test is thus more robust.